[feat] 루트 레이아웃 컴포넌트를 서버 컴포넌트로 전환 및 메타 데이터 추가
- client 컴포넌트를 래퍼 컴포넌트로 감싸서 분리
