+++ app/layout.tsx
@@ -1,16 +1,8 @@
-// TODO: 추후 서버 컴포넌트로 변환
-"use client";
-
-// import type { Metadata } from "next";
+import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import { DevTools } from "jotai-devtools";
-
-import { usePathname } from "next/navigation";
-import { AuthProvider } from "@/features/auth/model/AuthProvider";
-import AuthButtons from "@/features/auth/ui/AuthButtons";
-import { Toaster } from "sonner";
 import Menu from "@/components/Menu";
+import ClientProvider from "./ClientProvider";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -22,33 +14,26 @@ const geistMono = Geist_Mono({
   subsets: ["latin"],
 });
 
-// export const metadata: Metadata = {
-//   title: "Create Next App",
-//   description: "Generated by create next app",
-// };
+export const metadata: Metadata = {
+  title: "역도인 ",
+  description: "역도 훈련 도우미",
+};
 
 export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const pathname = usePathname();
-  const isVisible = !pathname.includes("login") && !pathname.includes("signup");
-
   return (
     <html lang="en">
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased container mx-auto px-4 max-w-md`}
       >
-        <AuthProvider>
-          {isVisible && <AuthButtons />}
-
-          <DevTools />
+        <ClientProvider>
           {children}
-          <Toaster />
 
           <Menu />
-        </AuthProvider>
+        </ClientProvider>
       </body>
     </html>
   );
